The user can delete all his activity history

+++ screens/UserScreen.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from "react"
-import { ActivityIndicator, Alert, Image, StyleSheet, Text, TouchableOpacity, View } from "react-native"
+import { ActivityIndicator, Alert, StyleSheet, Text, TouchableOpacity, View } from "react-native"
 import FontAwesome from '@expo/vector-icons/FontAwesome'
 
 import RecentSearch from "../components/RecentSearch"
@@ -7,7 +7,6 @@ import { SearchHistoryItem } from "../types/SearchHistoryItem"
 import { deleteSearchHistory, deleteHistoryItem, getHistory } from "../services/storageUtils"
 import { COLORS, SIZES } from "../constants"
 import { WatchedVideoItem } from "../types/WatchedVideoItem"
-import { SwipeListView } from "react-native-swipe-list-view"
 import RecentWatched from "../components/RecentWatched"
 import NoHistoryMsg from "../components/NoHistoryMsg"
 
@@ -19,8 +18,6 @@ export default function UserScreen() {
 
     useEffect(() => {
         fetchUserHistory()
-        console.log(JSON.stringify(recentWatched));
-
     }, [])
 
     const handleRefresh = () => {
@@ -30,9 +27,24 @@ export default function UserScreen() {
         setRefreshing(false)
     }
 
-    const handleDeleteAll = async () => {
-        await deleteSearchHistory()
-        fetchUserHistory()
+    const handleDeleteAll = () => {
+        Alert.alert(
+            "Delete All",
+            "Are you sure you want to delete all history?",
+            [{
+                text: "No",
+                style: "cancel"
+            },
+            {
+                text: "Yes",
+                onPress: async () => {
+                    await deleteSearchHistory()
+                    fetchUserHistory()
+                }
+            }
+            ],
+            { cancelable: false }
+        )
     }
 
     const handleDeleteEntity = async (type: string, value: string) => {
